% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatr.R
\name{gen_current}
\alias{gen_current}
\title{Get current value of a generator object}
\usage{
gen_current(gen)
}
\arguments{
\item{gen}{object of class \code{generator}}
}
\value{
The "current" value of a generator object. Will be a scalar value
 with type identical to that of the input generator.
}
\description{
Get current value of a generator object
}
\examples{
gen <- gen_init(1L:10L)
gen_get_current(gen)
}
