% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatr.R
\name{gen_init_empty}
\alias{gen_init_empty}
\title{Initialize an empty generator object}
\usage{
gen_init_empty(gen_type = c("integer", "numeric", "character", "logical"))
}
\arguments{
\item{gen_type}{char string, indicating the data type of the generator
object. Must be one of "integer", "numeric", "character", or "logical".}
}
\value{
object of class \code{generator}
}
\description{
Initialize an empty generator object
}
\examples{
gen <- gen_init(gen_type = "integer")
}
