% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatr.R
\name{gen_next}
\alias{gen_next}
\title{Get next value of a generator object}
\usage{
gen_next(gen)
}
\arguments{
\item{gen}{object of class \code{generator}}
}
\value{
The "next" value of a generator object. Will be a scalar value with
 type identical to that of the input generator.
}
\description{
Get next value of a generator object
}
\examples{
gen <- gen_init(1L:10L)
gen_next(gen)
}
